fix: extract changes in a new file to avoid actual download flow
